Restructure code and folders
