+++ src/components/DonutChart/DonutChart.js
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import PropTypes from 'prop-types'
 
 import { VictoryPie } from 'victory'
-import { mapApiTransportationTypeToDonutChartLabel } from '../services/mappings/mapApiTransportationTypeToDonutChartLabel'
+import { mapApiTransportationTypeToDonutChartLabel } from './mapApiTransportationTypeToDonutChartLabel'
 
 export default function DonutChart({ footprintData }) {
   const colors = [
