+++ src/components/DonutChart/mapApiTransportationTypeToDonutChartLabel.js
