+++ src/components/DonutChart/mapApiTransportationTypeToDonutChartLabel.test.js
