+++ src/pages/HomePage/HomePage.js
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import AuthUserContext from '../../components/auth/AuthUserContext'
 import Navigation from '../../components/Navigation/Navigation'
 import SumCarbonFootPrint from '../../components/SumCarbonFootprint/SumCarbonFootprint'
-import useDeviceDetect from '../../components/services/useDeviceDetect'
+import useDeviceDetect from '../../services/useDeviceDetect'
 import * as ROUTES from '../../constants/routes'
 import settings from '../../images/settings.svg'
 import homeIcon from '../../images/sustainable.svg'
