+++ src/services/calculateTotalFootprintSum.test.js
