+++ src/services/mappings/mapFootprintPerTransportTypeToDonutChartData.js
