+++ src/services/mappings/mapFootprintPerTransportTypeToDonutChartData.test.js
